Use two timers per cycle instead of a ticker
The previous implementation was in practice producing an output
frequency closer to 1.3Hz rather than the expected 10Hz. Moving to a
minimum pulse width of 1 millisecond actually improved the frequency to
just under 1Hz, which is much closer to the expected range.

This change removes the ticker, counter, and comparison loop in favor of
two timers: one that fires at the end of the 'mark' phase of the pulse,
and another that fires at the end of the 'space'. The measured ouput in
experiments with this implementation has been 10Hz +- .2Hz.